Saturn Sizing and Shadow
Fixed Saturn's Size to be a bit bigger than the default planet included in the demo scene.  Also added more detail to the rings by increasing the "Sort Distance" field.  The shadow from the planet onto the rings now works, I just need to fix the shadow on the back of the planet.  Also, the cloud/atmo-spheres don't seem to work properly and only cover half of the planet.  Overall lighting of the scene looks off for some reason, so I might try to fix that next.